Split epics demo mover and sensor into their own files
